+++ aats/schemas/governance.py
@@ -1,5 +1,7 @@
 from __future__ import annotations
 
+from decimal import Decimal
+
 from pydantic import Field
 
 from aats.schemas.common import SchemaBase
@@ -15,7 +17,7 @@ class PolicyDecision(SchemaBase):
     requires_human_approval: bool
     allowed_symbols: list[str]
     allowed_execution_styles: list[str]
-    max_notional_override: float | None = None
+    max_notional_override: Decimal | None = None
     forced_degrade_mode: str | None = None
     rejection_reasons: list[str] = Field(default_factory=list)
 
@@ -24,8 +26,8 @@ class RiskDecision(SchemaBase):
     decision_id: str
     approved: bool
     modified: bool
-    capped_target_position_qty: float
-    capped_target_notional: float | None = None
+    capped_target_position_qty: Decimal
+    capped_target_notional: Decimal | None = None
     current_open_order_count: int = 0
     constraints_applied: list[str] = Field(default_factory=list)
     risk_score: float
